# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Finsible - Personal Finance Management API
  description: |
    RESTful API for personal finance management including transaction tracking,
    account management, categorization, and multi-currency support.
    
    ## Authentication
    
    All endpoints except authentication endpoints require a valid JWT token.
    
    - **Web clients**: JWT is stored in an HttpOnly cookie
    - **Mobile clients**: JWT is returned in the response body and should be sent via Authorization header
    
    ## Response Format
    
    All responses follow a standard format:
    ```json
    {
      "message": "Description",
      "success": true/false,
      "timestamp": "2025-01-31 12:00:00",
      "data": { ... }
    }
    ```
  version: 0.0.1-SNAPSHOT
  contact:
    name: Finsible Support
    email: maitrymakwana196@gmail.com
servers:
  - url: http://localhost:9090
    description: Development server
  - url: https://api.finsible.org
    description: Production server
...
paths:
...
  /supported-currencies:
    get:
      tags:
        - Currencies
      summary: List supported currencies
      responses:
        '200':
          description: Currencies retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyListResponse'
    post:
      tags:
        - Currencies
      summary: Add a supported currency (Admin only)
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyRequest'
      responses:
        '200':
          description: Currency added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyResponse'
  /supported-currencies/{code}:
    get:
      tags:
        - Currencies
      summary: Get currency by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: ISO 4217 currency code
      responses:
        '200':
          description: Currency retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyResponse'
  /supported-currencies/{id}:
    delete:
      tags:
        - Currencies
      summary: Remove a supported currency (Admin only)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Currency removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt_token
      description: JWT token in HttpOnly cookie
  schemas:
    BaseResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    CurrencyRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 3
        name:
          type: string
          maxLength: 100
        symbol:
          type: string
          maxLength: 10
      required:
        - code
        - name
        - symbol
    CurrencyResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CurrencyData'
    CurrencyData:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
    CurrencyListResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CurrencyData'
...